requirements.txt
requests==2.22.0
gitpython==3.0.8
PTable==0.9.2